refact: reorganize types
