+++ src/types/apiPatterns.ts
@@ -1,0 +1,9 @@
+export interface SuccessResponse<T> {
+  timestamp: Date;
+  data: T;
+}
+
+export interface FailedResponse<T> {
+  error: number;
+  message: T;
+}
