--- src/types/apiRequests.ts
@@ -1,5 +1,0 @@
-export interface TeacherAuth {
-  token: string;
-  expiration: string;
-  userId: string;
-}
