+++ src/types/apiResponses.ts
@@ -1,9 +1,5 @@
-export interface SuccessResponse<T> {
-  timestamp: Date;
-  data: T;
-}
-
-export interface FailedResponse<T> {
-  error: number;
-  message: T;
+export interface TeacherAuth {
+  token: string;
+  expiration: string;
+  userId: string;
 }
